requirements.txt
requests
grpcio
grpcio-tools
google-pasta